data_ingestion/requirements.txt
apache-airflow-providers-google
pyarrow
psycopg2-binary
sqlalchemy